+++ CHANGELOG.rst
@@ -3,6 +3,11 @@ Version history
 
 This library adheres to `Semantic Versioning <http://semver.org/>`_.
 
+**UNRELEASED**
+
+- Argument type annotations are no longer unioned with the types of their default values, except in
+  the case of ``None`` as the default value (although PEP 484 still recommends against this)
+
 **2.2.0** (2018-07-08)
 
 - Fixed compatibility with Python 3.7